Move parser to ast module
