+++ src/vm/contracts.rs
@@ -3,7 +3,7 @@ use vm::representations::{SymbolicExpression};
 use vm::errors::{InterpreterResult as Result};
 use vm::callables::CallableType;
 use vm::contexts::{Environment, LocalContext, ContractContext, GlobalContext};
-use vm::parser;
+use vm::ast;
 use vm::types::QualifiedContractIdentifier;
 
 #[derive(Serialize, Deserialize)]
@@ -15,7 +15,7 @@ pub struct Contract {
 //          will probably be removed soon.
 impl Contract {
     pub fn initialize (contract_identifier: QualifiedContractIdentifier, contract: &str, global_context: &mut GlobalContext) -> Result<Contract> {
-        let parsed: Vec<_> = parser::parse(contract)?;
+        let parsed: Vec<_> = ast::parse(contract)?;
         let mut contract_context = ContractContext::new(contract_identifier);
 
         eval_all(&parsed, &mut contract_context, global_context)?;
